refactor: promote embedder-swap to its own lib module
The folder never met the action layer's bar — it persists through its own engine rather than mediating a store write — and it imports nothing from lib/actions, so it was already a leaf wearing the wrong address; as a sibling module it also retires the require-cycle workaround in per-turn-retrieval, whose three lazily-imported symbols were all embedder-swap's. withTurnAdmission joins the barrel because the turns now reach it across a boundary: it takes the admission lock rather than assuming it, so it is the safe-wrapper class the barrel already exports, not an engine primitive.
