+++ lib/actions/index.ts
@@ -5,38 +5,6 @@ export {
 } from './classifier/deps'
 export { runClassifierNow } from './classifier/run-now'
 export { defineAction } from './define-action'
-// The raw engine primitives (startSwap / resumeSwap / cancelSwap / reindexStory /
-// relabelModel) are deliberately absent: each requires the caller to already hold
-// the per-story single-flight lock, and the safe entry points are the *Story*
-// wrappers that take it via runExclusive. They stay reachable inside
-// `./embedder-swap` only, so the unsafe pair is never the more obvious import.
-export {
-  buildDrainController,
-  cancelStorySwap,
-  composeRetrievalEmbedDeps,
-  kickStoryDrain,
-  countStoryEmbeddableRows,
-  embedClassifierDescriptions,
-  refreshEmbeddingStatus,
-  reindexStoryNow,
-  relabelStory,
-  RelabelBlockedError,
-  RelabelDimMismatchError,
-  resolveDrainConfig,
-  resolveStorySwapConfig,
-  resumeStorySwap,
-  setDrainKickSink,
-  setDrainStatusSink,
-  startStorySwap,
-  SwapBusyError,
-  SwapConfigError,
-  SwapInProgressError,
-  SwapMarkerChangedError,
-  SwapNotInProgressError,
-  SwapStoryMissingError,
-  type StoryEmbedderActionRejection,
-  type SwapCancelOutcome,
-} from './embedder-swap'
 export { applyDeltaAction } from './delta/apply-delta-action'
 export { applyUndoPayload, computeUndoPayload } from './delta/delta-encoding'
 export { __resetRegistrationGuard, registerAllDomains } from './delta/registrations'
