+++ lib/embedder-swap/app-deps.test.ts
