+++ lib/embedder-swap/app-deps.ts
