+++ lib/embedder-swap/engine.ts
