+++ lib/pipeline/definitions/per-turn-retrieval.test.ts
@@ -1,7 +1,7 @@
 import type { DatabaseSync } from 'node:sqlite'
 
 import { eq } from 'drizzle-orm'
-import { afterEach, beforeAll, beforeEach, describe, expect, it, vi } from 'vitest'
+import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 
 import {
   APP_SETTINGS_DEFAULTS,
@@ -58,7 +58,7 @@ vi.mock('@/lib/retrieval', async (importOriginal) => {
   return { ...actual, runRetrieval: runRetrievalMock }
 })
 
-vi.mock('@/lib/actions', async (importOriginal) => {
+vi.mock('@/lib/embedder-swap', async (importOriginal) => {
   const actual = await importOriginal<Record<string, unknown>>()
   return { ...actual, refreshEmbeddingStatus: refreshEmbeddingStatusMock }
 })
@@ -342,12 +342,6 @@ function lastParams(): RetrievalParams {
   if (!call) throw new Error('runRetrieval was never called')
   return call[1] as RetrievalParams
 }
-
-// The phase imports @/lib/actions lazily to break a require cycle, so nothing
-// resolves that barrel until the first test does — inside a 5s timeout. Preload.
-beforeAll(async () => {
-  await import('@/lib/actions')
-}, 60_000)
 
 beforeEach(() => {
   vi.restoreAllMocks()
